Refactoring Previsions for History
