+++ app/src/main/java/com/example/tempoapp/ui/History/HistoryFragment.kt
@@ -1,4 +1,4 @@
-package com.example.tempoapp.ui.Previsions
+package com.example.tempoapp.ui.History
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,11 +7,11 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import com.example.tempoapp.databinding.FragmentPrevisionsBinding
+import com.example.tempoapp.databinding.FragmentHistoryBinding
 
-class PrevisionsFragment : Fragment() {
+class HistoryFragment : Fragment() {
 
-    private var _binding: FragmentPrevisionsBinding? = null
+    private var _binding: FragmentHistoryBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -22,14 +22,14 @@ class PrevisionsFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        val previsionsViewModel =
-            ViewModelProvider(this).get(PrevisionsViewModel::class.java)
+        val historyViewModel =
+            ViewModelProvider(this).get(HistoryViewModel::class.java)
 
-        _binding = FragmentPrevisionsBinding.inflate(inflater, container, false)
+        _binding = FragmentHistoryBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
         val textView: TextView = binding.textGallery
-        previsionsViewModel.text.observe(viewLifecycleOwner) {
+        historyViewModel.text.observe(viewLifecycleOwner) {
             textView.text = it
         }
         return root
