+++ app/src/main/res/layout/fragment_home.xml
@@ -14,6 +14,7 @@
         android:layout_marginStart="8dp"
         android:layout_marginTop="8dp"
         android:layout_marginEnd="8dp"
+        android:fontFamily="@font/raleway"
         android:textAlignment="center"
         android:textColor="#FFFFFF"
         android:textSize="20sp"
